the bot scroll for 1280 x 1080
Everything that happens from this bot is at your own risk once you activate it only you are in control
Or you download it and change the code in it and all you do with it is just messing around. I don't support the creation of bots that violate certain laws. It's simply a good way to learn, but to use several libraries at the same time, don't misuse it. It can simply help you understand how to build an application.